re-organized file structure (src/pages)
