+++ src/pages/TeamHome/TeamHome.tsx
@@ -1,7 +1,7 @@
 import { IonContent, IonPage, IonTitle } from '@ionic/react';
 import WidgetBot from '@widgetbot/react-embed';
 import NavBar from '../../components/NavBar';
-import './TeamHome.scss';
+import './teamHome.scss';
 
 const TeamHome: React.FC = () => {
   return (
