+++ src/pages/old/Tab1.tsx
@@ -1,9 +1,7 @@
 import { IonContent, IonPage, IonTitle } from '@ionic/react';
-import ExploreContainer from '../components/ExploreContainer';
-import NavBar from '../components/NavBar';
+import ExploreContainer from '../../components/ExploreContainer';
+import NavBar from '../../components/NavBar';
 import './Tab1.css';
-// sample data for team leaderboard & individual leaderboard (uncomment when using) //
-// import { TeamData, IndividualData } from './SampleData';
 
 const Tab1: React.FC = () => {
   return (
