+++ src/pages/old/Tab2.tsx
@@ -1,6 +1,6 @@
 import { IonContent, IonPage, IonTitle } from '@ionic/react';
-import ExploreContainer from '../components/ExploreContainer';
-import NavBar from '../components/NavBar';
+import ExploreContainer from '../../components/ExploreContainer';
+import NavBar from '../../components/NavBar';
 import './Tab2.css';
 
 const Tab2: React.FC = () => {
